pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tutor-stack"
version = "0.1.0"
description = "Tutor Stack platform - A comprehensive tutoring system"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Ahmed Sarhan", email = "ahmed.sarhan@example.com" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pyjwt[crypto]==2.10.1",
    "tutor-stack-content @ git+https://github.com/AhmedSarhanDL/tutor-stack-content.git",
    "tutor-stack-assessment @ git+https://github.com/AhmedSarhanDL/tutor-stack-assessment.git",
    "tutor-stack-notifier @ git+https://github.com/AhmedSarhanDL/tutor-stack-notifier.git",
    "tutor-stack-chat @ git+https://github.com/AhmedSarhanDL/tutor-stack-tutor_chat.git",
    "tutor-stack-auth @ git+https://github.com/AhmedSarhanDL/tutor-stack-auth.git"
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.5",
    "httpx==0.27.0",
    "pytest-cov==4.1.0",
    "black==24.2.0",
    "isort==5.13.2",
    "flake8==7.0.0",
    "mypy==1.8.0"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = []

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100 